requirements.txt
asyncio
aiohttp
disnake==2.5.0
disnake-jishaku
motor
dnspython
emoji
hmtai
textwrap3
easy_pil
bs4
git+https://github.com/Leyla-developers/Lavalink.py.git
PyNaCl
SmallGoogleTranslator
python-dotenv
genshin
psutil
git+https://github.com/Leyla-developers/leyla-justify.git
git+https://github.com/MagM1go/DiscordActivity.git
wikipedia
git+https://github.com/boticord/blurplefier.git
git+https://github.com/MagM1go/logg.git
redis>=4.2.0rc1
flask[async]
nest_asyncio